added commandline option --benchmark NxN. fixes to last commit, cleanup, smaller data structures in task queue, signal wakes all threads fix

+++ veejay-current/veejay-server/libvje/effects/motionblur.c
@@ -105,7 +105,7 @@ void motionblur_apply( VJFrame *frame, int width, int height, int n) {
 			n_motion_frames = 1;
 		}
 	}
-	int t = num_threaded_tasks();
+	uint8_t t = num_threaded_tasks();
 	if( t <= 0 ) t = 1;
 	if(n_motion_frames >= ( n * t ) ) {
 		n_motion_frames = 0;
